added test cases and changed the return type to double

+++ src/main/java/co/interleap/courses/tdd/StatementGenerator.java
@@ -1,7 +1,7 @@
 package co.interleap.courses.tdd;
 
 public class StatementGenerator {
-    int getFare(int distance,int minutes)
+    double getFare(double distance,int minutes)
     {
         return distance*10+minutes*1;
     }
